requirements.txt
requests==2.32.5
html2text==2025.4.15
openai==2.8.1
python-dotenv==1.2.1
tiktoken==0.12.0